feat: referencer ownage
