+++ src/integration-tests/test-cases/json-schema-meta-schema.json
@@ -1,7 +1,3 @@
 {
-  "title": "lesSchema",
-  "type": "array",
-  "items": {
-    "$ref": "https://raw.githubusercontent.com/json-schema-tools/meta-schema/master/meta-schema.json"
-  }
+  "$ref": "https://raw.githubusercontent.com/json-schema-tools/meta-schema/master/meta-schema.json"
 }
